Support HLS (in development)
1. Why is this change made?
Prior to this change, when you open `.ts` files, the page will crash directly. This change will fix it, and try to support the display of HLS videos, eg. `.ts`, `.m3u8`. However, DPlayer is not working in my case, and it's still under investigation.

2. What is changed?
Add `.ts` in supported file extensions and add a handler.

3. What else is concerning?
Video playback is not working (page won't crash now).