Add a button to make the download easier.
1. Why is the change made?
If there are hundreds of files we want to download, a better way is to provide the download link directly instead of the preview page.
This change adds a new button in all the folderViews that makes it possible to switch between the download mode and the preview mode.

2. What has been changed?
Introduced a new url param: `direct`. If it's in the url, all the urls in the web page will be changed: folders will add `direct` param, and files will add `raw` param. Also added a button which simply adds/removes the `direct` param from the request.

4. What else is concerning?
A more delicate way is to use Microsoft's API for batch download. May dive into that later.